requirements.txt
pandas>=1.5,<3
openpyxl>=3.1
PyYAML>=6.0
requests>=2.31
